Better decouples multiphase logic from standard one
